+++ include/fx2timer.h
@@ -17,10 +17,11 @@
 #ifndef FX2_TIMER_H
 #define FX2_TIMER_H
 
+#include <stdint.h>
 #include "fx2types.h"
 
 void fx2_timer0_isr();
-void fx2_setup_timer0(WORD us);
+void fx2_setup_timer0(uint32_t us);
 
 extern void timer0_callback();
